Add files via upload
Marketing mix stands as a widely utilized concept in the execution of marketing strategies. It encompasses various facets within a comprehensive marketing plan, with a central focus on the four Ps of marketing: product, price, place, and promotion. As a data scientist, EDA and hypothesis testing has been conducted to enhance the comprehension of the diverse factors influencing customer acquisition.